Reorganized exu | Added comments to exu
